Moved over Grid Space, started getting Grid ready
Grid Space is moved over, with the main change being that it combines the old Grid Space and Piece Space. Inventory is also moved over (with rotating moved over to Piece), though I'm not sure if it'll stay in the final product like that. Next up is porting over Grid (with some temp assets already in the repo) and making sure everything still works.