Simple loss condition added
When you grab a piece, it makes sure there's a legal place for it to go (in any rotation), and if not (or if it fails to place garbage 500 times) it says you lose--just a console message for now. Also fixed an oversight with the preview pane and connected the powerup bar for Player 2.